Creacion y terminado de registro de pedido sin pruebas de unidad, se modidfico la base de datos, se agrego la llave primaria a orderProduct y se actualizo proyecto DataAccess

+++ ItaliaPizza/Logic/ProductLogic.cs
@@ -34,7 +34,7 @@ namespace Logic
 							Image = ImageLogic.ConvertToBitMapImage(product.picture),
 							Preparation = product.preparation,
 							IdRecipe = product.idRecipe,
-							Quantity = product.quantity	
+							Quantity = (int)product.quantity	
 						};
 
 						productsObtained.Add(recoverProduct);
